convert community cloud and docs page
